requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.1
django-crispy-forms==1.9.0
django-filter==2.2.0
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
mysqlclient==1.4.5
Pillow==7.1.1
Pygments==2.6.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
